Upload file for creation date bash script
This script takes the EXIF date when the picture was digitized and makes it the title. This helps when figuring out timelines of specific pictures due to some OS's using the their own metadata like 'last opened' or 'created' which most of the time it does not accurately represent the creation date. to be honest this annoyed me so much that it is why I created this script. In my mind who would want the date you uploaded a file as the date it uses as created? Sure the file was put there on that date and time, but it does not actually represent the date created.